auth  fixes for testing
