+++ utilities/shared.js
@@ -1,8 +1,6 @@
 import DatabaseController from "../database/mongo/controller.js"
 import Project from '../classes/Project/Project.js'
-import { findPageById } from '../page/index.js'
-
-export { findPageById }
+import Page from '../classes/Page/Page.js'
 
 /**
  * Check if the supplied input is valid JSON or not.
@@ -98,3 +96,39 @@ export const getProjectById = async (projectId, res) => {
       layer.pages.some(p => p.id.split('/').pop() === pageId.split('/').pop())
       )
    }
+
+// Find a page by ID (moved from page/index.js)
+export async function findPageById(pageId, projectId) {
+  if (pageId?.startsWith(process.env.RERUMIDPREFIX)) {
+    return fetch(pageId).then(res => res.json())
+  }
+  const projectData = (await getProjectById(projectId))?.data
+  if (!projectData) {
+    const error = new Error(`Project with ID '${projectId}' not found`)
+    error.status = 404
+    throw error
+  }
+  const layerContainingPage = projectData.layers.find(layer =>
+    layer.pages.some(p => p.id.split('/').pop() === pageId.split('/').pop())
+  )
+
+  if (!layerContainingPage) {
+    const error = new Error(`Layer containing page with ID '${pageId}' not found in project '${projectId}'`)
+    error.status = 404
+    throw error
+  }
+
+  const pageIndex = layerContainingPage.pages.findIndex(p => p.id.split('/').pop() === pageId.split('/').pop())
+
+  if (pageIndex < 0) {
+    const error = new Error(`Page with ID '${pageId}' not found in project '${projectId}'`)
+    error.status = 404
+    throw error
+  }
+
+  const page = layerContainingPage.pages[pageIndex]
+  page.prev = layerContainingPage.pages[pageIndex - 1] ?? null
+  page.next = layerContainingPage.pages[pageIndex + 1] ?? null
+
+  return new Page(layerContainingPage.id, page)
+}
